Added Calendar and AM logic

+++ db/migrate/20200324135356_create_recipe_details.rb
@@ -3,7 +3,7 @@ class CreateRecipeDetails < ActiveRecord::Migration[6.0]
     create_table :recipe_details do |t|
       t.integer :ingredient_id
       t.integer :recipe_id
-      t.integer :servings
+      t.float :servings
       t.string :serving_unit
 
       t.timestamps
